Update clawing of bears when a player claws
Whenever a player claws, a timer of 1000ms is set and every 50ms it will
decrease with 50ms until it becomes 0. During this time the bear's
'claw' will be true and it will be nil when it reaches 0. This is needed
to temporarily set a 'clawing' class on bears in the game.